pyproject.toml
[tool.black]
line-length = 88
skip-string-normalization = true
preview = true
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.mypy_cache
    | \.tox
    | \.venv
    | migrations
  )/
)
'''

[tool.isort]
line_length = 88
skip_glob = ['*/migrations/*']
# Use vertical hanging indents
multi_line_output = 3
use_parentheses = true
# Co-mingle "import x" and "from x import y"
force_sort_within_sections = true
# Co-mingle constnats, classes, and modules
order_by_type = false
case_sensitive = false
# Sort . ahead of .. (primarily to appease flake8-import-order)
reverse_relative = true
profile = 'black'

[tool.poetry]
name = "teste"
version = "0.1.0"
description = ""
authors = ["Luca Lacerda <lucalacerda1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
dependency-injector = "^4.41.0"
requests = "^2.31.0"
uvicorn = "^0.24.0.post1"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.11.0"
ipdb = "^0.13.13"
pylama = "^8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
